refacotor and add parseDbf
