+++ web/src/classic/core/mantle/data/shapefile/shapefile.ts
@@ -9,6 +9,7 @@ import { processGeoJSON } from "../geojson";
 import { f, FetchOptions } from "../utils";
 
 import { parseDbf } from "./parseDbf";
+import { parseShp } from "./parseShp";
 
 async function binaryAjax(url: string, type?: string): Promise<Buffer | string | false> {
   const fullUrl = type ? `${url}.${type}` : url;
@@ -30,391 +31,6 @@ async function binaryAjax(url: string, type?: string): Promise<Buffer | string |
       return false;
     }
     throw e;
-  }
-}
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-async function unzip(buffer: Buffer): Promise<{ [key: string]: Buffer | string }> {
-  const zip = new JSZip();
-  await zip.loadAsync(buffer);
-  const files = zip.file(/.+/);
-  const out: { [key: string]: Buffer | string } = {};
-
-  await Promise.all(
-    files.map(async a => {
-      let result;
-      if (a.name.slice(-3).toLowerCase() === "shp" || a.name.slice(-3).toLowerCase() === "dbf") {
-        result = await a.async("nodebuffer");
-      } else {
-        result = await a.async("text");
-      }
-      out[a.name] = result;
-    }),
-  );
-
-  return out;
-}
-
-// function isClockWise(array: number[][]): {
-//   ring: number[][];
-//   clockWise: boolean;
-//   bbox: number[];
-//   children: number[][][];
-// } {
-//   // Implementation of isClockWise function
-//   // ...
-// }
-
-// function contains(outer: { bbox: number[] }, inner: { bbox: number[] }): boolean {
-//   // Implementation of contains function
-//   // ...
-// }
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function handleRings(rings: number[][][]) {
-  // Implementation of handleRings function
-  // ...
-}
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-async function parseZip(buffer: Buffer | ArrayBuffer, whiteList?: string[]) {
-  // Implementation of parseZip function
-  // ...
-}
-
-function parseShp(buffer: Buffer, trans?: proj4.Converter): GeoJSONFeature<GeometryObject>[] {
-  const headers = parseHeader(buffer);
-  const parseFunc = getParseFunction(headers.shpCode, trans);
-  const rows = getRows(buffer, headers, parseFunc);
-  return rows;
-}
-
-function parseHeader(buffer: Buffer): {
-  length: number;
-  version: number;
-  shpCode: number;
-  bbox: number[];
-} {
-  const view = buffer.subarray(0, 100);
-  return {
-    length: view.readInt32BE(6 << 2) << 1,
-    version: view.readInt32LE(7 << 2),
-    shpCode: view.readInt32LE(8 << 2),
-    bbox: [
-      view.readDoubleLE(9 << 2),
-      view.readDoubleLE(11 << 2),
-      view.readDoubleLE(13 << 2),
-      view.readDoubleLE(15 << 2),
-    ],
-  };
-}
-
-function getParseFunction(
-  shpCode: number,
-  trans?: proj4.Converter,
-): (data: Buffer) => GeoJSONFeature<GeometryObject> | null {
-  const num = shpCode > 20 ? shpCode - 20 : shpCode;
-  const shpFuncObj: { [key: number]: keyof typeof parseFunctions } = {
-    1: "parsePoint",
-    3: "parsePolyline",
-    5: "parsePolygon",
-    8: "parseMultiPoint",
-    11: "parseZPoint",
-    13: "parseZPolyline",
-    15: "parseZPolygon",
-    18: "parseZMultiPoint",
-  };
-  const funcName = shpFuncObj[num];
-  if (!funcName) {
-    throw new Error("Unsupported shape type");
-  }
-  const parseCoord = makeParseCoord(trans);
-  return (data: Buffer) => parseFunctions[funcName](data, parseCoord);
-}
-
-function getRows(
-  buffer: Buffer,
-  headers: ReturnType<typeof parseHeader>,
-  parseFunc: ReturnType<typeof getParseFunction>,
-): GeoJSONFeature<GeometryObject>[] {
-  const rows: GeoJSONFeature<GeometryObject>[] = [];
-  let offset = 100;
-  const bufferLength = buffer.byteLength;
-  while (offset + 8 <= bufferLength) {
-    const record = getRow(buffer, offset, bufferLength);
-    if (!record) {
-      break;
-    }
-    offset += 8 + record.len;
-    const feature = parseFunc(record.data);
-    if (feature) {
-      rows.push(feature);
-    }
-  }
-  return rows;
-}
-
-function getRow(
-  buffer: Buffer,
-  offset: number,
-  bufferLength: number,
-):
-  | {
-      id: number;
-      len: number;
-      data: Buffer;
-      type: number;
-    }
-  | undefined {
-  const view = buffer.subarray(offset, offset + 12);
-  const len = view.readInt32BE(4) << 1;
-  const id = view.readInt32BE(0);
-  if (len === 0) {
-    return {
-      id,
-      len,
-      data: Buffer.alloc(0),
-      type: 0,
-    };
-  }
-  if (offset + len + 8 > bufferLength) {
-    return undefined;
-  }
-  return {
-    id,
-    len,
-    data: buffer.subarray(offset + 12, offset + len + 8),
-    type: view.readInt32LE(8),
-  };
-}
-
-const parseFunctions = {
-  parsePoint(
-    data: Buffer,
-    parseCoord: ReturnType<typeof makeParseCoord>,
-  ): GeoJSONFeature<GeometryObject> {
-    return {
-      type: "Feature",
-      geometry: {
-        type: "Point",
-        coordinates: parseCoord(data, 0),
-      },
-      properties: {},
-    };
-  },
-
-  parseZPoint(
-    data: Buffer,
-    parseCoord: ReturnType<typeof makeParseCoord>,
-  ): GeoJSONFeature<GeometryObject> {
-    const pointXY = parseFunctions.parsePoint(data, parseCoord);
-    (pointXY.geometry as any).coordinates.push(data.readDoubleLE(16));
-    return pointXY;
-  },
-
-  parsePointArray(
-    data: Buffer,
-    offset: number,
-    num: number,
-    parseCoord: ReturnType<typeof makeParseCoord>,
-  ): number[][] {
-    const points: number[][] = [];
-    for (let i = 0; i < num; i++) {
-      points.push(parseCoord(data, offset));
-      offset += 16;
-    }
-    return points;
-  },
-
-  parseZPointArray(
-    data: Buffer,
-    zOffset: number,
-    num: number,
-    coordinates: number[][],
-  ): number[][] {
-    for (let i = 0; i < num; i++) {
-      coordinates[i].push(data.readDoubleLE(zOffset));
-      zOffset += 8;
-    }
-    return coordinates;
-  },
-
-  parseArrayGroup(
-    data: Buffer,
-    offset: number,
-    partOffset: number,
-    num: number,
-    tot: number,
-    parseCoord: ReturnType<typeof makeParseCoord>,
-  ): number[][][] {
-    const groups: number[][][] = [];
-    let pointOffset = offset;
-    for (let i = 0; i < num; i++) {
-      const pointNum =
-        i === num - 1
-          ? tot - data.readInt32LE(partOffset)
-          : data.readInt32LE(partOffset + 4) - data.readInt32LE(partOffset);
-      partOffset += 4;
-      if (pointNum === 0) {
-        continue;
-      }
-      groups.push(parseFunctions.parsePointArray(data, pointOffset, pointNum, parseCoord));
-      pointOffset += pointNum << 4;
-    }
-    return groups;
-  },
-
-  parseZArrayGroup(
-    data: Buffer,
-    zOffset: number,
-    num: number,
-    coordinates: number[][][],
-  ): number[][][] {
-    for (let i = 0; i < num; i++) {
-      coordinates[i] = parseFunctions.parseZPointArray(
-        data,
-        zOffset,
-        coordinates[i].length,
-        coordinates[i],
-      );
-      zOffset += coordinates[i].length << 3;
-    }
-    return coordinates;
-  },
-
-  parseMultiPoint(
-    data: Buffer,
-    parseCoord: ReturnType<typeof makeParseCoord>,
-  ): GeoJSONFeature<GeometryObject> {
-    const num = data.readInt32LE(32);
-    if (num === 0) {
-      return {
-        type: "Feature",
-        geometry: {
-          type: "MultiPoint",
-          coordinates: [],
-        },
-        properties: {},
-      };
-    }
-    const bounds = [parseCoord(data, 0), parseCoord(data, 16)];
-    const pointOffset = 36;
-    const coordinates = parseFunctions.parsePointArray(data, pointOffset, num, parseCoord);
-    return {
-      type: "Feature",
-      geometry: {
-        type: "MultiPoint",
-        coordinates,
-      },
-      properties: {},
-      bbox: [bounds[0][0], bounds[0][1], bounds[1][0], bounds[1][1]],
-    };
-  },
-
-  parseZMultiPoint(
-    data: Buffer,
-    parseCoord: ReturnType<typeof makeParseCoord>,
-  ): GeoJSONFeature<GeometryObject> {
-    const feature = parseFunctions.parseMultiPoint(data, parseCoord);
-    const num = (feature.geometry as any).coordinates.length;
-    const zOffset = 52 + (num << 4);
-    (feature.geometry as any).coordinates = parseFunctions.parseZPointArray(
-      data,
-      zOffset,
-      num,
-      (feature.geometry as any).coordinates,
-    );
-    return feature;
-  },
-
-  parsePolyline(
-    data: Buffer,
-    parseCoord: ReturnType<typeof makeParseCoord>,
-  ): GeoJSONFeature<GeometryObject> {
-    const numParts = data.readInt32LE(32);
-    if (numParts === 0) {
-      return {
-        type: "Feature",
-        geometry: {
-          type: "LineString",
-          coordinates: [],
-        },
-        properties: {},
-      };
-    }
-    const bounds = [parseCoord(data, 0), parseCoord(data, 16)];
-    const num = data.readInt32LE(36);
-    const partOffset = 40;
-    const pointOffset = 40 + (numParts << 2);
-    const parts = parseFunctions.parseArrayGroup(
-      data,
-      pointOffset,
-      partOffset,
-      numParts,
-      num,
-      parseCoord,
-    );
-    return {
-      type: "Feature",
-      geometry: {
-        type: "MultiLineString",
-        coordinates: parts,
-      },
-      properties: {},
-      bbox: [bounds[0][0], bounds[0][1], bounds[1][0], bounds[1][1]],
-    };
-  },
-
-  parseZPolyline(
-    data: Buffer,
-    parseCoord: ReturnType<typeof makeParseCoord>,
-  ): GeoJSONFeature<GeometryObject> {
-    const feature = parseFunctions.parsePolyline(data, parseCoord);
-    const numParts = data.readInt32LE(32);
-    const num = data.readInt32LE(36);
-    const zOffset = 56 + (num << 4) + (numParts << 2);
-    (feature.geometry as any).coordinates = parseFunctions.parseZArrayGroup(
-      data,
-      zOffset,
-      numParts,
-      (feature.geometry as any).coordinates,
-    );
-    return feature;
-  },
-
-  parsePolygon(
-    data: Buffer,
-    parseCoord: ReturnType<typeof makeParseCoord>,
-  ): GeoJSONFeature<GeometryObject> {
-    const feature = parseFunctions.parsePolyline(data, parseCoord);
-    (feature.geometry as any).type = "MultiPolygon";
-    (feature.geometry as any).coordinates = handleRings((feature.geometry as any).coordinates);
-    return feature;
-  },
-
-  parseZPolygon(
-    data: Buffer,
-    parseCoord: ReturnType<typeof makeParseCoord>,
-  ): GeoJSONFeature<GeometryObject> {
-    const feature = parseFunctions.parseZPolyline(data, parseCoord);
-    (feature.geometry as any).type = "MultiPolygon";
-    (feature.geometry as any).coordinates = handleRings((feature.geometry as any).coordinates);
-    return feature;
-  },
-};
-
-function makeParseCoord(trans?: proj4.Converter): (data: Buffer, offset: number) => number[] {
-  if (trans) {
-    return (data: Buffer, offset: number) => {
-      const x = data.readDoubleLE(offset);
-      const y = data.readDoubleLE(offset + 8);
-      return trans.inverse([x, y]);
-    };
-  } else {
-    return (data: Buffer, offset: number) => [
-      data.readDoubleLE(offset),
-      data.readDoubleLE(offset + 8),
-    ];
   }
 }
